refactor(create): rename 'world cardinal create' command to just 'world create' (#31)
Closes: WORLD-732

## Overview

Rename `world cardinal create` command to `world create`

## Brief Changelog

- Move create.go from cardinal package to root package
- Add create command to the root command

## Testing and Verifying

Test `world create` command is working properly
