+++ cmd/world/cardinal/cardinal.go
@@ -9,10 +9,10 @@ import (
 
 func init() {
 	// Register subcommands - `world cardinal [subcommand]`
-	BaseCmd.AddCommand(createCmd, startCmd, devCmd, restartCmd, purgeCmd, stopCmd)
+	BaseCmd.AddCommand(startCmd, devCmd, restartCmd, purgeCmd, stopCmd)
 
 	// Add --debug flag
-	logger.AddLogFlag(createCmd, startCmd, devCmd, restartCmd, purgeCmd, stopCmd)
+	logger.AddLogFlag(startCmd, devCmd, restartCmd, purgeCmd, stopCmd)
 }
 
 // BaseCmd is the base command for the cardinal subcommand
